+++ src/main/java/listener/AttentionListener.java
@@ -1,0 +1,9 @@
+package listener;
+
+import event.AttentionEvent;
+
+public interface AttentionListener {
+
+    void onAttentionEvent(AttentionEvent event);
+
+}