+++ src/main/java/listener/EEGPowerListener.java
@@ -1,0 +1,9 @@
+package listener;
+
+import event.EEGPowerEvent;
+
+public interface EEGPowerListener {
+
+    void onEEGPowerEvent(EEGPowerEvent event);
+
+}