+++ src/main/java/listener/MeditationListener.java
@@ -1,0 +1,9 @@
+package listener;
+
+import event.MeditationEvent;
+
+public interface MeditationListener {
+
+    void onMeditationEvent(MeditationEvent event);
+
+}