+++ src/main/java/listener/PoorSignalListener.java
@@ -1,0 +1,9 @@
+package listener;
+
+import event.PoorSignalEvent;
+
+public interface PoorSignalListener {
+
+    void onPoorSignalEvent(PoorSignalEvent event);
+
+}