+++ src/main/java/listener/RawWaveListener.java
@@ -1,0 +1,9 @@
+package listener;
+
+import event.RawWaveEvent;
+
+public interface RawWaveListener {
+
+    void onRawWaveEvent(RawWaveEvent event);
+
+}