+++ src/main/java/listener/StrengthListener.java
@@ -1,0 +1,9 @@
+package listener;
+
+import event.StrengthEvent;
+
+public interface StrengthListener {
+
+    void onStrengthEvent(StrengthEvent event);
+
+}