+++ src/main/kotlin/event/AttentionEvent.kt
@@ -1,0 +1,4 @@
+package event
+
+class AttentionEvent(val value: Int): DeviceEvent {
+}