+++ src/main/kotlin/event/MeditationEvent.kt
@@ -1,0 +1,4 @@
+package event
+
+class MeditationEvent(val value: Int): DeviceEvent {
+}