+++ src/main/kotlin/event/PoorSignalEvent.kt
@@ -1,0 +1,4 @@
+package event
+
+class PoorSignalEvent(val quality: Int): DeviceEvent {
+}