+++ src/main/kotlin/event/RawWaveEvent.kt
@@ -1,0 +1,4 @@
+package event
+
+class RawWaveEvent(val value: Int): DeviceEvent {
+}