+++ CHANGELOG.md
@@ -8,11 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - EEG raw data viewer
 
+## [0.2.1] - 2021-09-27
+### Added
+- DataReceiver, receive device data and decode it
+- possible event and listener
+
 ## [0.2.0] - 2021-09-22
 ### Added
 - Using fazecast's jSerialComm library communicate with mind wave mobile device
 ### Changed
-- Update kotlin version
 - Change kt file locate from java folder to kotlin folder
 ### Removed
 - ThinkGear because thinkgear DLL is 32bits version